pyproject.toml
[project]
name = "astro-events"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.4"
dependencies = [
    "pyephem>=9.99",
]
